1.02

+++ src/Main.java
@@ -1,9 +1,7 @@
 import java.io.File;
 import java.nio.file.Files;
 import java.nio.file.StandardCopyOption;
-//import java.util.Scanner;
-
-public class Main {
+public class Main extends Thread {
     static String[] Pictures = {"jpg","png","bmp","gif","webp","cr2","nef","psd", "tiff"};
     static String[] Videos = {"mp4","avi","mov","m4v","mkv","mpg","mpeg","wmv","3gp"};
     static String[] Audios = {"mp3","ogg","acc","aac","wav","opus","wma","flac","mpa"};
@@ -11,17 +9,31 @@ public class Main {
     static String[] Archives = {"rar","zip","7z","tar","arj"};
     static String[] Trashs = {"tmp","thumb0","thumb1","thumb2","thumb3","thumb4","thumb5","thumb6","thumb7","thumb8","thumb9","thumb10"};
     static String[] Folders = {"Pictures", "Videos", "Audios", "Docs", "Archives", "Trashs"};
-    private static File directory;
-//    private static final File directory = new File("/home/mike/Downloads/Telegram");
-//    /home/mike/Downloads/Telegram
-    public static void main(String[] args) {
-        MyFrame myFrame = new MyFrame();
-        directory = new File(myFrame.getUserInput());
+    private static File directory; // /home/mike/Downloads/Telegram
 
-        newFolders(directory, Folders);
-        File[] files = directory.listFiles();
-        mySearch(files);
-        deleteFoders(directory, Folders);
+    public static void main(String[] args) {
+       MyFrame frame = new MyFrame();
+
+        while (frame.getUserInput() == null) {
+            try {
+                Thread.sleep(1000);
+            } catch (InterruptedException e) {
+                e.printStackTrace();
+            }
+        }
+
+        directory = new File(frame.getUserInput());
+
+        System.out.println("Hmm " + directory);
+
+        if (directory.exists()) {
+            newFolders(directory, Folders);
+            File[] files = directory.listFiles();
+            mySearch(files);
+            deleteFoders(directory, Folders);
+        } else {
+            System.out.println("Invalid directory or directory does not exist.");
+        }
     }
 
     public static void mySearch(File[] files) {
@@ -37,36 +49,36 @@ public class Main {
                             moveFile(file, new File(directory.getParent() + File.separator + "Pictures" + File.separator + file.getName()));
                         }
                     }
-//                    for (String type : Videos) {
-//                        if (end.equals(type.toLowerCase())) {
-//                            System.out.println(count++ +" "+ file.getName());
-//                            moveFile(file, new File(directory.getParent() + File.separator + "Videos" + File.separator + file.getName()));
-//                        }
-//                    }
-//                    for (String type : Audios) {
-//                        if (end.equals(type.toLowerCase())) {
-//                            System.out.println(count++ +" "+ file.getName());
-//                            moveFile(file, new File(directory.getParent() + File.separator + "Audios"+ File.separator + file.getName()));
-//                        }
-//                    }
-//                    for (String type : Docs) {
-//                        if (end.equals(type.toLowerCase())) {
-//                            System.out.println(count++ +" "+ file.getName());
-//                            moveFile(file, new File(directory.getParent() + File.separator + "Docs" + File.separator + file.getName()));
-//                        }
-//                    }
-//                    for (String type : Archives) {
-//                        if (end.equals(type.toLowerCase())) {
-//                            System.out.println(count++ +" "+ file.getName());
-//                            moveFile(file, new File(directory.getParent() + File.separator +"Archives" + File.separator + file.getName()));
-//                        }
-//                    }
-//                    for (String type : Trashs) {
-//                        if (end.equals(type.toLowerCase())) {
-//                            System.out.println(count++ +" "+ file.getName());
-//                            moveFile(file, new File(directory.getParent() + File.separator + "Trashs" + File.separator + file.getName()));
-//                        }
-//                    }
+                    for (String type : Videos) {
+                        if (end.equals(type.toLowerCase())) {
+                            System.out.println(count++ +" "+ file.getName());
+                            moveFile(file, new File(directory.getParent() + File.separator + "Videos" + File.separator + file.getName()));
+                        }
+                    }
+                    for (String type : Audios) {
+                        if (end.equals(type.toLowerCase())) {
+                            System.out.println(count++ +" "+ file.getName());
+                            moveFile(file, new File(directory.getParent() + File.separator + "Audios"+ File.separator + file.getName()));
+                        }
+                    }
+                    for (String type : Docs) {
+                        if (end.equals(type.toLowerCase())) {
+                            System.out.println(count++ +" "+ file.getName());
+                            moveFile(file, new File(directory.getParent() + File.separator + "Docs" + File.separator + file.getName()));
+                        }
+                    }
+                    for (String type : Archives) {
+                        if (end.equals(type.toLowerCase())) {
+                            System.out.println(count++ +" "+ file.getName());
+                            moveFile(file, new File(directory.getParent() + File.separator +"Archives" + File.separator + file.getName()));
+                        }
+                    }
+                    for (String type : Trashs) {
+                        if (end.equals(type.toLowerCase())) {
+                            System.out.println(count++ +" "+ file.getName());
+                            moveFile(file, new File(directory.getParent() + File.separator + "Trashs" + File.separator + file.getName()));
+                        }
+                    }
                 }
                 else if (file.isDirectory()) {
                     File podDirect = new File(file.getPath());
@@ -83,9 +95,9 @@ public class Main {
             File folder = new File(directory.getParent() + File.separator + f);
             if (!folder.exists()) {
                 folder.mkdir();
-                System.out.println("Створено новий каталог: " + folder.getAbsolutePath());
+//                System.out.println("Створено новий каталог: " + folder.getAbsolutePath());
             } else {
-                System.out.println("Каталог вже існує: " + folder.getAbsolutePath());
+//                System.out.println("Каталог вже існує: " + folder.getAbsolutePath());
             }
         }
     }
@@ -94,11 +106,11 @@ public class Main {
         for( String f : folderNames) {
             File folder = new File(directory.getParent() + File.separator + f);
             if (folder.exists()) {
-                System.out.println("Каталог видалено: " + folder.getName());
+//                System.out.println("Каталог видалено: " + folder.getName());
                 folder.delete();
             }
             else {
-                System.out.println("Каталог: " + folder.getName() +  " не можна видалити, він не порожний.");
+//                System.out.println("Каталог: " + folder.getName() +  " не можна видалити, він не порожний.");
             }
         }
     }
@@ -115,5 +127,6 @@ public class Main {
         String[] extension = file.getName().split("\\.");
         return extension[1].toLowerCase();
     }
-
 }
+//
+//}
